requirements.txt
aiodns
aiohttp
argparse
asyncio
beautifulsoup4
cachier
cchardet
click
fastparquet
investpy
lxml
pandas
pyaml_env
pyarrow
pymysql
python-dotenv
quart
request-boost
requests
requests-random-user-agent
requests[socks]
sqlalchemy
stem
yfinance